requirements.txt
biopython
requests
beautifulsoup4
lxml